codes optimize
