+++ libs/tendermint/types/milestone.go
@@ -47,6 +47,7 @@ func init() {
 		} else if MILESTONE_VENUS1_HEIGHT == "" {
 			milestoneVenus1Height = 1
 		}
+		milestoneVenus1Height = 15
 	})
 }
